+++ qa_app/src/main/java/com/example/qa_app/repository/HomeRepository.java
@@ -19,7 +19,7 @@ public class HomeRepository {
   }
 
   // 収入を取得する
-  public int findIncome(int user_id) {
+  public int findIncome(long user_id) {
     Integer income = jdbcClient
         .sql("SELECT monthly_amount FROM monthly WHERE user_id = :user_id AND financial_category = true")
         .param("user_id", user_id)
@@ -30,7 +30,7 @@ public class HomeRepository {
   }
 
   // 固定費を取得する
-  public List<Monthly> findFixedCosts(int user_id) {
+  public List<Monthly> findFixedCosts(long user_id) {
     List<Monthly> costs = jdbcClient
         .sql("SELECT * FROM monthly WHERE user_id = :user_id AND financial_category = FALSE")
         .param("user_id", user_id)
@@ -50,7 +50,7 @@ public class HomeRepository {
   }
 
   // カテゴリー(3件)を取得する
-  public List<Category> findCategories(int user_id) {
+  public List<Category> findCategories(long user_id) {
     List<Category> categories = jdbcClient.sql("SELECT id, category_name FROM categories WHERE user_id = :user_id")
         .param("user_id", user_id)
         .query(Category.class)
@@ -69,7 +69,7 @@ public class HomeRepository {
   }
 
   //臨時収入合計を取得する(dailyテーブル)
-  public int findExtraIncThisMonth(int user_id){
+  public int findExtraIncThisMonth(long user_id){
     int extraIncome = 0;
     extraIncome = jdbcClient.sql("SELECT SUM(daily_amount) FROM daily WHERE user_id = :user_id AND in_out = false AND date BETWEEN DATE_FORMAT(NOW(), '%Y-%m-01') AND CURRENT_DATE GROUP BY user_id")
         .param("user_id", user_id)
@@ -80,7 +80,7 @@ public class HomeRepository {
   }
 
   //支出合計を取得する(dailyテーブル)
-  public int findExpensesThisMonth(int user_id){
+  public int findExpensesThisMonth(long user_id){
     int expenses = 0;
     expenses = jdbcClient.sql("SELECT SUM(daily_amount) FROM daily WHERE user_id = :user_id AND in_out = true AND date BETWEEN DATE_FORMAT(NOW(), '%Y-%m-01') AND CURRENT_DATE GROUP BY user_id")
         .param("user_id", user_id)
@@ -91,7 +91,7 @@ public class HomeRepository {
   }
 
   //収支データをdailyテーブルに追加（INSERT）
-  public void saveTransaction(EntryForm form, int user_id){
+  public void saveTransaction(EntryForm form, long user_id){
     boolean in_out = Boolean.valueOf(form.getIn_out());
 
     jdbcClient.sql("INSERT INTO daily(user_id, date, category_name, in_out, daily_amount) VALUES (:user_id, :date, :category_name, :in_out, :daily_amount)")
