+++ qa_app/src/main/java/com/example/qa_app/service/MyPageService.java
@@ -26,36 +26,36 @@ public class MyPageService extends BaseService {
     //収入の表示、今月使用可能金額の予算・固定費合計の計算はBaseServiceを継承
 
     //固定費の項目名、金額を表示する
-    public List<Monthly> showFixedCosts(int user_id) {
+    public List<Monthly> showFixedCosts(long user_id) {
         return myPageRepository.findFixedCosts(user_id);
     }
 
     //収入を変更する
-    public void changeIncome(int user_id, int income) {
+    public void changeIncome(long user_id, int income) {
         myPageRepository.changeIncome(user_id, income);
     }
 
     //カテゴリーを表示する
-    public List<CategoryEditForm> showCategoriesMypage(int user_id){
+    public List<CategoryEditForm> showCategoriesMypage(long user_id){
         return myPageRepository.findCategories(user_id);
     }
     
     //カテゴリを変更する
-    public void changeCategories(int user_id, List<CategoryEditForm> list) {
+    public void changeCategories(long user_id, List<CategoryEditForm> list) {
         for (CategoryEditForm c : list) {
             myPageRepository.changeCategories(user_id, c);
         }
     }
 
     //固定費の項目を削除する
-    public void deleteCost(int user_id, Long id, Monthly monthly) {
+    public void deleteCost(long user_id, Long id, Monthly monthly) {
         if (monthly.isFinCategory() == false) {
             myPageRepository.deleteCost(user_id, id);
         }
     }
 
     //固定費の項目を追加する
-    public void addCost(int user_id, MonthlyEditForm form) {
+    public void addCost(long user_id, MonthlyEditForm form) {
         if (form.getFixedName() != null) {
             myPageRepository.addCost(user_id, form);
         }
